LayoutTests/storage/websql/open-database-creation-callback-expected.txt

CONSOLE MESSAGE: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: unable to open database, version mismatch, '1.0' does not match the currentVersion of ''
CONSOLE MESSAGE: Web SQL is deprecated. Please use IndexedDB instead.
CONSOLE MESSAGE: Web SQL is deprecated. Please use IndexedDB instead.
This test tests openDatabase()'s creation callback.
